requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
pillow~=10.3.0
tqdm~=4.66.2
beautifulsoup4~=4.12.3
matplotlib~=3.8.4
QtPy~=2.4.1
PyYAML~=6.0.1
requests~=2.31.0
paddleseg~=2.7.0
setuptools~=69.2.0
scipy~=1.13.0
sklearn~=0.0
scikit-learn~=1.4.2
pandas~=2.2.2
threadpoolctl~=3.5.0
filelock~=3.14.0
six~=1.16.0
visualdl~=2.5.3
gradio~=4.28.3

scikit-image~=0.23.2
imageio~=2.34.1
